Vision/tutorial02_VGG/requirements.txt
tqdm
Pillow 
numpy 
torch 
tabulate